refactor: put organization input into form data and change order
