+++ web-app/src/app/(auth)/components/form/form-fields.tsx
@@ -17,20 +17,25 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import { OrganizationWithMembersCount } from "@/lib/db";
 import { FormData } from "@/lib/form";
+import { Organization } from "@/prisma/generated/client";
 
+import { OrganizationInput } from "./organization-input";
 import { AuthNamespace } from "./types";
 
 interface FormFieldsProps<T extends FieldValues> {
   form: UseFormReturn<T>;
   formData: FormData<T, AuthNamespace>;
   namespace: AuthNamespace;
+  organizations?: OrganizationWithMembersCount[] | undefined;
 }
 
 export function FormFields<T extends FieldValues>({
   form,
   formData,
   namespace,
+  organizations,
 }: FormFieldsProps<T>) {
   const t = useTranslations(namespace);
 
@@ -44,7 +49,12 @@ export function FormFields<T extends FieldValues>({
           render={({ field }) => (
             <FormItem>
               <FormControl>
-                <FormInput field={field} formDataItem={formDataItem} t={t} />
+                <FormInput
+                  field={field}
+                  formDataItem={formDataItem}
+                  t={t}
+                  organizations={organizations}
+                />
               </FormControl>
               <FormMessage />
             </FormItem>
@@ -59,12 +69,14 @@ interface FormInputProps<T extends FieldValues> {
   field: ControllerRenderProps<T, Path<T>>;
   formDataItem: FormData<T, AuthNamespace>[number];
   t: IntlTranslation<AuthNamespace>;
+  organizations?: OrganizationWithMembersCount[] | undefined;
 }
 
 function FormInput<T extends FieldValues>({
   field,
   formDataItem,
   t,
+  organizations,
 }: FormInputProps<T>) {
   const { type, labelKey, name, placeholderKey } = formDataItem;
 
@@ -80,6 +92,24 @@ function FormInput<T extends FieldValues>({
           {labelKey && t(labelKey)}
         </Label>
       </div>
+    );
+  }
+
+  if (name === "organizationId" && !!organizations) {
+    const organizationId = field.value;
+    const organization = organizations.find(
+      (organization) => organization.id === organizationId,
+    );
+    const handleOrganizationChange = (organization: Organization) => {
+      field.onChange(organization.id);
+    };
+
+    return (
+      <OrganizationInput
+        organizations={organizations}
+        value={organization}
+        onChange={handleOrganizationChange}
+      />
     );
   }
