+++ web-app/src/app/(auth)/register/data.ts
@@ -32,6 +32,9 @@ const signUpFormData: FormData<SignUpFormSchemaType, "Auth.Pages.SignUp.Form"> =
       placeholderKey: "Fields.Name.placeholder",
     },
     {
+      name: "organizationId",
+    },
+    {
       name: "email",
       placeholderKey: "Fields.Email.placeholder",
     },
